+++ src/components/forms/MultiValueInput/MultiValueInput.test.tsx
@@ -87,6 +87,23 @@ describe('MultiValueInput', () => {
       expect(mockValueAdd).toHaveBeenCalledTimes(1);
       expect(mockValueAdd).toHaveBeenCalledWith(pastedValue, pastedValue);
     });
+    it('should add a comma separated values list', () => {
+      const pastedValue = ['value1', 'value2', 'value3'];
+      render(<MultiValueInput valuesDelimiter="," onValueAdd={mockValueAdd} />);
+
+      screen.getByRole('textbox').focus();
+      userEvent.paste(
+        pastedValue.join(','),
+        {
+          clipboardData: {
+            getData: () => pastedValue.join(','),
+          },
+        } as unknown, // as I really don't need other properties of this object
+      );
+
+      expect(mockValueAdd).toHaveBeenCalledTimes(1);
+      expect(mockValueAdd).toHaveBeenCalledWith(pastedValue, pastedValue);
+    });
     it('should add new value on blur', async () => {
       const inputValue = 'value';
       render(<MultiValueInput onValueAdd={mockValueAdd} />);