+++ src/components/typographyLegacy/Text/Text.stories.tsx
@@ -92,8 +92,6 @@ export const TextHeading5: Story = () => (
   </Stack>
 );
 
-SmallText.storyName = 'Small Text (DEPRECATED)';
-
 export const PrimaryText: Story = () => (
   <>
     <Text>{lipsum}</Text>